+++ mods/include/mods/init/init.h
@@ -33,6 +33,7 @@ void init_death();
 void init_options();
 void init_chat();
 void init_bucket();
+void init_cake();
 void init_home();
 
 #ifdef __cplusplus