+++ mods/src/init/init.c
@@ -32,6 +32,7 @@ __attribute__((constructor)) static void init(int argc, char *argv[]) {
     init_options();
     init_chat();
     init_bucket();
+    init_cake();
     init_home();
 #ifndef MCPI_SERVER_MODE
     init_benchmark(argc, argv);